# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/actualites": {
            "get": {
                "tags": [
                    "Actualites"
                ],
                "summary": "Get list of actualites",
                "description": "Returns list of actualites",
                "operationId": "getActualitesList",
                "parameters": [
                    {
                        "name": "actualite",
                        "in": "path",
                        "description": "actualite id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Actualites"
                ],
                "summary": "modify actualites",
                "description": "return modified actualite",
                "operationId": "getActualitesList",
                "parameters": [
                    {
                        "name": "actualite",
                        "in": "path",
                        "description": "actualite id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}